+++ config/filesystems.php
@@ -68,6 +68,13 @@ return [
             'report' => false,
         ],
 
+        'site-promotions' => [
+            'driver' => 'local',
+            'root' => storage_path('app/site-promotions'),
+            'throw' => false,
+            'report' => false,
+        ],
+
         's3' => [
             'driver' => 's3',
             'key' => env('AWS_ACCESS_KEY_ID'),